+++ cpp/models/secir_vaccine/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_library(secir_vaccine
-    age_group.h
     infection_state.h
     analyze_result.h
     analyze_result.cpp
