--- cpp/models/secir_vaccine/age_group.h
@@ -1,44 +1,0 @@
-/* 
-* Copyright (C) 2020-2021 German Aerospace Center (DLR-SC)
-*
-* Authors: Jan Kleinert, Daniel Abele
-*
-* Contact: Martin J. Kuehn <Martin.Kuehn@DLR.de>
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-#ifndef AGEGROUP_H
-#define AGEGROUP_H
-
-#include "memilio/utils/index.h"
-
-namespace mio
-{
-namespace vaccinated
-{
-
-    /**
- * @brief The AgeGroup struct is used as a dynamically
- * sized tag for all age dependent categories
- */
-    struct AgeGroup : public Index<AgeGroup> {
-        AgeGroup(size_t val)
-            : Index<AgeGroup>(val)
-        {
-        }
-    };
-
-} // namespace vaccinated
-} // namespace mio
-
-#endif
